+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
 ### [1.1.3](https://github.com/textcompare/random-addresses-generator/compare/v1.1.2...v1.1.3) (2025-03-01)
+### Added
+- Generate addresses for particular cities by their names.
 
 ### 1.1.2 (2025-02-20)